requirements.txt
pycaw
pywin32
pystray
pillow